move methods to correct adapter
