+++ infrastructure/postgres-adapter/src/main/java/onlydust/com/marketplace/api/postgres/adapter/PostgresRewardStatusAdapter.java
@@ -2,17 +2,12 @@ package onlydust.com.marketplace.api.postgres.adapter;
 
 import lombok.AllArgsConstructor;
 import lombok.NonNull;
-import onlydust.com.marketplace.accounting.domain.model.ProjectId;
 import onlydust.com.marketplace.accounting.domain.model.RewardId;
 import onlydust.com.marketplace.accounting.domain.model.RewardStatusData;
 import onlydust.com.marketplace.accounting.domain.model.billingprofile.BillingProfile;
-import onlydust.com.marketplace.accounting.domain.model.user.UserId;
 import onlydust.com.marketplace.accounting.domain.port.out.RewardStatusStorage;
-import onlydust.com.marketplace.api.postgres.adapter.entity.write.RewardEntity;
 import onlydust.com.marketplace.api.postgres.adapter.entity.write.RewardStatusDataEntity;
-import onlydust.com.marketplace.api.postgres.adapter.repository.RewardRepository;
 import onlydust.com.marketplace.api.postgres.adapter.repository.RewardStatusRepository;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -20,7 +15,6 @@ import java.util.Optional;
 @AllArgsConstructor
 public class PostgresRewardStatusAdapter implements RewardStatusStorage {
     private final RewardStatusRepository rewardStatusRepository;
-    private final RewardRepository rewardRepository;
 
     @Override
     public void save(RewardStatusData rewardStatusData) {
@@ -58,19 +52,5 @@ public class PostgresRewardStatusAdapter implements RewardStatusStorage {
                 .stream()
                 .map(RewardStatusDataEntity::toRewardStatus)
                 .toList();
-    }
-
-    @Override
-    @Transactional
-    public void updateBillingProfileForRecipientUserIdAndProjectId(BillingProfile.Id billingProfileId, UserId userId, ProjectId projectId) {
-        rewardRepository.updateBillingProfileForRecipientUserIdAndProjectId(billingProfileId.value(), userId.value(), projectId.value());
-    }
-
-    @Override
-    @Transactional
-    public List<RewardId> removeBillingProfile(BillingProfile.Id billingProfileId) {
-        final var rewardIds = rewardRepository.getRewardIdsToBeRemovedFromBillingProfile(billingProfileId.value()).stream().map(RewardEntity::id).toList();
-        rewardRepository.removeBillingProfileIdOf(rewardIds);
-        return rewardIds.stream().map(RewardId::of).toList();
     }
 }
